Fixed types and formatting

+++ src/rw_top_trumps.h
@@ -10,7 +10,7 @@ extern "C" {
 void evaluate_rw_top_trumps(char *suite_name, size_t number_of_objectives, size_t function,
     size_t instance, size_t dimension, const double *x, double *y);
 
-void rw_top_trumps_bounds(size_t instance, int m, double *min, double *max);
+void rw_top_trumps_bounds(size_t instance, size_t m, double *min, double *max);
 
 void rw_top_trumps_test(void);
